fix: eliminate all remaining panics and unsafe patterns across workspace
- consensus: fix is_finalized chain linking, remove .unwrap() in finality/proposer/validator
- crypto: replace .try_into().unwrap() with copy_from_slice in poseidon2 and threshold
- pvm: encode_immediate/encode_mem_immediate return Option instead of assert! panic
- pvm/asm: propagate encoding errors via asm_encode_imm helper
- mempool: remove dead hash_cache, BlockDecryptor::new returns Result, fix .unwrap()s
- tx: fix u128→u64 truncation in pipeline U256 conversions
- otic: replace bare .unwrap() with .expect() messages, fix parser pattern match

+++ crates/tx/src/pipeline.rs
@@ -257,10 +257,10 @@ fn execute_in_pvm(
     let ctx = ExecutionContext {
         caller: tx.from,
         self_address: tx.to,
-        call_value: U256::from(tx.value as u64),
+        call_value: U256::from(tx.value),
         block_number: block_ctx.height,
         timestamp: block_ctx.timestamp,
-        gas_price: U256::from(block_ctx.base_fee as u64),
+        gas_price: U256::from(block_ctx.base_fee),
         tx_nonce: tx.nonce,
         tx_gas_limit: tx.gas_limit,
         tx_hash: U256::ZERO,
